Switched data type.

+++ Utilities/test_rknn.py
@@ -37,7 +37,7 @@ orig_h, orig_w = img.shape[:2]
 orig = img.copy()
 
 img_resized, scale, pad_x, pad_y = letterbox(img, (640, 640))
-img_input = np.expand_dims(img_resized, 0).astype(np.float32)
+img_input = np.expand_dims(img_resized, 0).astype(np.uint8)
 img_input = np.ascontiguousarray(img_input)
 
 # -----------------
